+++ Commands/utilsCommands.py
@@ -1,6 +1,9 @@
 #!/usr/bin/env python
 # -*- coding: utf-8 -*-
-
+# A library that provides functionality to the @randomutils_bot
+# Copyright (C) 2017-2018
+# Javier Gines Sanchez <software@javisite.com>
+#
 
 import logging												## System module
 log = logging.getLogger(__name__)
@@ -8,7 +11,7 @@ log = logging.getLogger(__name__)
 from subprocess import call									## System module
 import os													## System module
 from platform import system									## System module
-from sys import exc_info
+from sys import exc_info									## System module
 from random import randint									## System module
 
 import Functions.basicData as bd							## Own module