+++ Functions/basicData.py
@@ -1,11 +1,16 @@
 #!/usr/bin/env python
 # -*- coding: utf-8 -*-
+# A library that provides functionality to the @randomutils_bot
+# Copyright (C) 2017-2018
+# Javier Gines Sanchez <software@javisite.com>
+#
 
 import logging												## System module
 log = logging.getLogger(__name__)
-from sys import exc_info
 
-import Functions.message as ms
+from sys import exc_info									## System module
+
+import Functions.message as ms								## Own module
 
 
 ## Debug chat_id
@@ -26,16 +31,12 @@ def startWithCommand(bot, update, args=['']):
 	global chat_id
 	global user_id
 
-	if update.message == None:
-		message = update.edited_message
-	else:
-		message = update.message
+	message = update.edited_message if update.message == None else update.message
 
 	username = message.from_user.name
 	chat_id = message.chat.id
 	user_id = message.from_user.id
 
-	log.info(str(message.photo))
 	if message.forward_from != None:
 		textToLog = "Forward Message" + ' --> ' + username + " (chat_id:" + str(chat_id) + " , user_id:"+ str(user_id) + ")"
 	elif message.text != None: